refactored tests, added read folder functionality

+++ read_fasta.h
@@ -6,4 +6,4 @@
 
 using namespace std;
 
-int read_fasta(string &filepath, vector<string> &chrnames, string &content, vector<unsigned long long> &chrpos);
+int read_fasta(string &filepath, vector<string> &chrnames, string &content, vector<uint64_t> &chrpos);
